DoubleML

Abstract base class for DoubleML models.

* Methods `$tune()` and `$fit()` which call internal methods for parameter tuning and estimation of the causal effects.

Public fields

data

(`DoubleMLData`)
Data object.

n_folds

(`integer(1)`)
Number of folds used for cross-fitting.

ml_learners

(`list()`)
List of learners as named in mlr3 and mlr3learners packages.

params

(`list()`)
List of parameters passed to learners.

dml_procedure

(`character(1)`)
Double Machine Learning algorithm used (either "dml1" or "dml2").

inf_model

(`character(1)`)
Score to be used for estimation.

subgroups

(`list()`)
Subgroups.

n_rep_cross_fit

(`integer(1)`)
Number of repetitions for cross-fitting.

se_reestimate

(`logical(1)`)
Flag for reestimation of standard error after cross-fitting.

param_set

(`list()`)
List with parameter set from mlr3tuning of class `ParamSet`.

tune_settings

(`list()`)
List with settings for parameter tuning passed to mlr3tuning.

Methods

Public methods

DoubleML$new()

DoubleML$fit()

DoubleML$bootstrap()

DoubleML$set_samples()

DoubleML$tune()

DoubleML$summary()

DoubleML$confint()

DoubleML$clone()

Method new()

Initialize instance. `DoubleML` is an abstract class and initialization is typically performed for inherited classes.

Usage

DoubleML$new(...)

Returns

coef (`numeric(1)`)
Final coefficient estimate.

Method fit()

Usage

DoubleML$fit(se_reestimate = FALSE)

Method bootstrap()

Usage

DoubleML$bootstrap(method = "normal", n_rep = 500)

Method set_samples()

Usage

DoubleML$set_samples(smpls)

Method tune()

Usage

DoubleML$tune()

Method summary()

Usage

DoubleML$summary(digits = max(3L, getOption("digits") - 3L))

Method confint()

Usage

DoubleML$confint(parm, level = 0.95, joint = FALSE)

Method clone()

The objects of this class are cloneable with this method.

Usage

DoubleML$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.